pyproject.toml
[tool.poetry]
name = "longbi"
version = "0.1.0"
description = ""
authors = ["jebi <aedara@uw.edu>"]
license = "MIT License"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pillow = "^10.0.0"
pytesseract = "^0.3.10"
numpy = "^1.25.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
